Changed Dependency Strategy to require packages already included
